refactor(polygon): move shader parsing code into gl mod
